Move to logger folder
